refactor: move canvas node display rules to entity
